improved default edit control width
git-svn-id: https://tuftsvue.svn.sourceforge.net/svnroot/tuftsvue/trunk@3280 d66e5487-c4fa-4dda-8c8e-ca23240363c0

+++ VUE2/src/edu/tufts/vue/ui/ConfigurationUI.java
@@ -153,7 +153,7 @@ public class ConfigurationUI extends javax.swing.JPanel
 					return;
 				}
 				
-				Integer numChars = new Integer(0);
+				Integer numChars = new Integer(15);
 				if (maxChars != null) {
 					try {
 						numChars = (new Integer(maxChars));
